hooks: rename to gate_hooks
