+++ core/gate_hooks/pcapng.cc
